functions refactoring: primary and secondary subpackages added
