+++ connectors/filesystem/abstract_file.py
@@ -3,23 +3,25 @@ from typing import Optional, Iterable, Union, Any
 import os
 
 try:  # Assume we're a sub-module in a package.
-    from utils import arguments as arg, dates as dt
+    from utils import arguments as arg
     from utils.decorators import deprecated_with_alternative
     from interfaces import (
         Context, Connector, ConnectorInterface, IterableStreamInterface, ItemType, StreamType,
         AUTO, Auto, AutoName, AutoCount, AutoBool, AutoConnector, OptionalFields,
     )
+    from functions.primary import dates as dt
     from connectors.abstract.leaf_connector import LeafConnector
     from connectors.mixin.streamable_mixin import StreamableMixin
     from streams.mixin.stream_builder_mixin import StreamBuilderMixin
     from streams.mixin.iterable_mixin import IterableStreamMixin
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
-    from ...utils import arguments as arg, dates as dt
+    from ...utils import arguments as arg
     from ...utils.decorators import deprecated_with_alternative
     from ...interfaces import (
         Context, Connector, ConnectorInterface, IterableStreamInterface, ItemType, StreamType,
         AUTO, Auto, AutoName, AutoCount, AutoBool, AutoConnector, OptionalFields,
     )
+    from ...functions.primary import dates as dt
     from ..abstract.leaf_connector import LeafConnector
     from ..mixin.streamable_mixin import StreamableMixin
     from ...streams.mixin.stream_builder_mixin import StreamBuilderMixin
@@ -202,9 +204,10 @@ class AbstractFile(LeafConnector, StreamableMixin, ABC):
     def get_count(self, allow_reopen: bool = True, allow_slow_gzip: bool = True, force: bool = False) -> Optional[int]:
         pass
 
-    def is_empty(self) -> bool:
-        count = self.get_count(allow_slow_gzip=False) or 0
-        return count <= 0
+    def is_empty(self) -> Optional[bool]:
+        count = self.get_count(allow_slow_gzip=False)
+        if count is not None:
+            return count <= 0
 
     def is_in_memory(self) -> bool:
         return False
