+++ functions/all_functions.py
@@ -1,13 +1,13 @@
 try:  # Assume we're a sub-module in a package.
-    from functions.basic_functions import (
+    from functions.secondary.basic_functions import (
         partial, const, defined, is_none, not_none, nonzero, equal, not_equal,
         at_least, more_than, safe_more_than, less_than, between, not_between, is_ordered,
         apply_dict,
     )
-    from functions.cast_functions import DICT_CAST_TYPES, cast, date, number, percent
-    from functions.numeric_functions import sign, diff, div, mult, sqrt
-    from functions.date_functions import int_to_date, date_to_int, round_date, next_date, date_range
-    from functions.array_functions import (
+    from functions.secondary.cast_functions import DICT_CAST_TYPES, cast, date, number, percent
+    from functions.secondary.numeric_functions import sign, diff, div, mult, sqrt
+    from functions.secondary.date_functions import int_to_date, date_to_int, round_date, next_date, date_range
+    from functions.secondary.array_functions import (
         is_in, not_in,
         elem_no, first, second, last,
         distinct, uniq, count_uniq, count,
@@ -15,24 +15,24 @@ try:  # Assume we're a sub-module in a package.
         values_not_none, defined_values, nonzero_values, numeric_values, shift_right,
         unfold_lists, top, hist, mean,
     )
-    from functions.aggregate_functions import avg, median, min, max, sum
-    from functions.pair_functions import shifted_func, pair_filter, pair_stat, corr
-    from functions.logic_functions import maybe, always, never
-    from functions.item_functions import (
+    from functions.secondary.aggregate_functions import avg, median, min, max, sum
+    from functions.secondary.pair_functions import shifted_func, pair_filter, pair_stat, corr
+    from functions.secondary.logic_functions import maybe, always, never
+    from functions.secondary.item_functions import (
         composite_key, value_by_key, values_by_keys, is_in_sample,
         same, merge_two_items, items_to_dict,
         json_loads, csv_loads, csv_reader,
     )
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
-    from .basic_functions import (
+    from .secondary.basic_functions import (
         partial, const, defined, is_none, not_none, nonzero, equal, not_equal,
         at_least, more_than, safe_more_than, less_than, between, not_between, is_ordered,
         apply_dict,
     )
-    from .cast_functions import DICT_CAST_TYPES, cast, date, number, percent
-    from .numeric_functions import sign, diff, div, mult, sqrt
-    from .date_functions import int_to_date, date_to_int, round_date, next_date, date_range
-    from .array_functions import (
+    from .secondary.cast_functions import DICT_CAST_TYPES, cast, date, number, percent
+    from .secondary.numeric_functions import sign, diff, div, mult, sqrt
+    from .secondary.date_functions import int_to_date, date_to_int, round_date, next_date, date_range
+    from .secondary.array_functions import (
         is_in, not_in,
         elem_no, first, second, last,
         distinct, uniq, count_uniq, count,
@@ -40,10 +40,10 @@ except ImportError:  # Apparently no higher-level package has been imported, fal
         values_not_none, defined_values, nonzero_values, numeric_values, shift_right,
         unfold_lists, top, hist, mean,
     )
-    from .aggregate_functions import avg, median, min, max, sum
-    from .pair_functions import shifted_func, pair_filter, pair_stat, corr
-    from .logic_functions import maybe, always, never
-    from .item_functions import (
+    from .secondary.aggregate_functions import avg, median, min, max, sum
+    from .secondary.pair_functions import shifted_func, pair_filter, pair_stat, corr
+    from .secondary.logic_functions import maybe, always, never
+    from .secondary.item_functions import (
         composite_key, value_by_key, values_by_keys, is_in_sample,
         same, merge_two_items, items_to_dict,
         json_loads, csv_loads, csv_reader,
