+++ functions/primary/dates.py
@@ -1,12 +1,12 @@
-from typing import Union, Optional, NoReturn
+from typing import Union, Optional
 from datetime import date, timedelta, datetime
 
 try:  # Assume we're a sub-module in a package.
     from utils import arguments as arg
     from utils.enum import DynamicEnum
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
-    from . import arguments as arg
-    from .enum import DynamicEnum
+    from ...utils import arguments as arg
+    from ...utils.enum import DynamicEnum
 
 PyDate = date
 IsoDate = str
@@ -66,7 +66,7 @@ def get_min_year() -> int:
     return _min_year
 
 
-def set_min_year(year: int) -> NoReturn:
+def set_min_year(year: int):
     global _min_year
     _min_year = year
