+++ functions/secondary/item_functions.py
@@ -9,7 +9,7 @@ try:  # Assume we're a sub-module in a package.
         items as it,
     )
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
-    from ..utils import (
+    from ...utils import (
         arguments as arg,
         selection as sf,
         items as it,
