+++ functions/secondary/readme.md
@@ -1,0 +1,8 @@
+## Secondary functions (meta-functions)
+
+### What is it
+
+Meta-functions are second-level functions
+ returning parametrized functions 
+ for using as mappers, selectors, filters 
+ in relevant [stream](../../streams/readme.md) methods.
