+++ items/flat_struct.py
@@ -11,7 +11,7 @@ try:  # Assume we're a sub-module in a package.
     from base.abstract.simple_data import SimpleDataWrapper
     from fields.advanced_field import AdvancedField
     from selection.abstract_expression import AbstractDescription
-    from functions import array_functions as fs
+    from functions.secondary import array_functions as fs
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
     from ..interfaces import (
         StructInterface, StructRowInterface, FieldInterface, SelectionLoggerInterface, ExtLogger,
