+++ series/simple/date_series.py
@@ -3,7 +3,7 @@ from typing import Iterable, Callable, Optional, Union
 
 try:  # Assume we're a sub-module in a package.
     from series import series_classes as sc
-    from utils import dates as dt
+    from functions.primary import dates as dt
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
     from .. import series_classes as sc
     from ...utils import dates as dt
