+++ streams/mixin/iterable_mixin.py
@@ -12,7 +12,7 @@ try:  # Assume we're a sub-module in a package.
         Source, ExtLogger, SelectionLogger, LoggingLevel,
         AUTO, Auto, AutoName, AutoCount, Count, OptionalFields, UniKey,
     )
-    from functions import item_functions as fs
+    from functions.secondary import item_functions as fs
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
     from ...utils import algo, arguments as arg
     from ...utils.external import pd, DataFrame, get_use_objects_for_output
